Reorganizacion de Frontend - y Backend

+++ AparcaYa/BackEnd-AparcaYA/src/main/java/com/exe/AparcaYA/Repository/ReservacionRepository.java
@@ -9,76 +9,56 @@ import org.springframework.stereotype.Repository;
 import java.time.LocalDateTime;
 import java.util.List;
 
+// ✅ CAMBIO #9: JpaRepository<Reservacion, Integer> → JpaRepository<Reservacion, Long>
+// La Entity Reservacion declara `private Long idReserva`, el tipo del Repository
+// debe coincidir. Usar Integer forzaba Math.toIntExact() en el Impl con riesgo
+// de ArithmeticException para IDs grandes.
 @Repository
-public interface ReservacionRepository extends JpaRepository<Reservacion, Integer> {
+public interface ReservacionRepository extends JpaRepository<Reservacion, Long> {
 
-    /**
-     * Obtiene todas las reservaciones de un cliente específico
-     */
-    List<Reservacion> findByClienteIdUsuario(Integer clienteId);
+    // ✅ CAMBIO #9: Integer clienteId → Long clienteId en todos los métodos
+    List<Reservacion> findByClienteIdUsuario(Long clienteId);
 
-    /**
-     * Obtiene reservaciones de un cliente por estado
-     */
-    List<Reservacion> findByClienteIdUsuarioAndEstado(Integer clienteId, String estado);
+    List<Reservacion> findByClienteIdUsuarioAndEstado(Long clienteId, String estado);
 
-    /**
-     * Obtiene reservaciones por estado
-     */
     List<Reservacion> findByEstado(String estado);
 
-    /**
-     * Obtiene reservaciones por sede
-     */
+    // ✅ CAMBIO #9: Integer sedeId → Long sedeId
     @Query("SELECT r FROM Reservacion r WHERE r.cupo.sede.idSede = :sedeId")
-    List<Reservacion> findByCupoSedeId(@Param("sedeId") Integer sedeId);
+    List<Reservacion> findByCupoSedeId(@Param("sedeId") Long sedeId);
 
-    /**
-     * Obtiene reservaciones por cupo
-     */
-    List<Reservacion> findByCupoIdCupo(Integer cupoId);
+    // ✅ CAMBIO #9: Integer cupoId → Long cupoId
+    List<Reservacion> findByCupoIdCupo(Long cupoId);
 
-    /**
-     * Obtiene reservaciones activas de un cupo en un rango de fechas
-     */
     @Query("SELECT r FROM Reservacion r WHERE r.cupo.idCupo = :cupoId " +
             "AND r.estado IN :estados " +
             "AND r.fechaInicio < :fechaFin " +
             "AND r.fechaFin > :fechaInicio")
     List<Reservacion> findConflictosHorario(
-            @Param("cupoId") Integer cupoId,
+            @Param("cupoId") Long cupoId,
             @Param("estados") List<String> estados,
             @Param("fechaInicio") LocalDateTime fechaInicio,
             @Param("fechaFin") LocalDateTime fechaFin
     );
 
-    /**
-     * Obtiene reservaciones de un cupo con estados específicos
-     */
-    List<Reservacion> findByCupoIdCupoAndEstadoIn(Integer cupoId, List<String> estados);
+    List<Reservacion> findByCupoIdCupoAndEstadoIn(Long cupoId, List<String> estados);
 
-    /**
-     * Obtiene reservaciones por vehículo
-     */
     List<Reservacion> findByVehiculoPlaca(String placa);
 
-    /**
-     * Obtiene reservaciones futuras de un cliente
-     */
+    // ✅ CAMBIO #9: Integer clienteId → Long clienteId
     @Query("SELECT r FROM Reservacion r WHERE r.cliente.idUsuario = :clienteId " +
             "AND r.fechaInicio > :fechaActual " +
             "ORDER BY r.fechaInicio ASC")
     List<Reservacion> findReservacionesFuturas(
-            @Param("clienteId") Integer clienteId,
+            @Param("clienteId") Long clienteId,
             @Param("fechaActual") LocalDateTime fechaActual
     );
 
-    /**
-     * Cuenta reservaciones activas de un cliente
-     */
+    // ✅ CAMBIO #9: Integer clienteId → Long clienteId
     @Query("SELECT COUNT(r) FROM Reservacion r WHERE r.cliente.idUsuario = :clienteId " +
             "AND r.estado = 'ACTIVA'")
-    long countReservacionesActivas(@Param("clienteId") Integer clienteId);
+    long countReservacionesActivas(@Param("clienteId") Long clienteId);
 
+    // Este método ya estaba correcto con Long — se mantiene igual
     List<Reservacion> findByCliente_IdUsuario(Long idUsuario);
 }
